Refactor image generation package to unify OpenAI and xAI client implementations; remove xAI-specific options and streamline API usage
